refactor: собрать все элементы товара в одной папке
