+++ src/components/Account/AccountBuyerOrdersList/OrderListItem/OrderListItem.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import './OrderListItem.scss';
-import ProductCardBasket from '../../../ProductCardBasket/ProductCardBasket';
+import ProductCardBasket from '../../../ProductElements/ProductCardBasket/ProductCardBasket';
 
 const OrderListItem = ({ order }) => {
   return <div className="order-list-item"></div>;
