+++ src/components/ProductElements/CharacteristicColor/CharacteristicColor.jsx
