+++ src/components/ProductElements/CharacteristicColor/CharacteristicColor.scss
@@ -1,4 +1,4 @@
-@import '../../styles/global';
+@import '../../../styles/global';
 
 .characteristic-color {
   display: flex;
