+++ src/components/ProductElements/CommentsBlock/CommentsBlock.scss
@@ -1,4 +1,4 @@
-@import '../../styles/global';
+@import '../../../styles/global';
 
 .comment {
   width: 403px;
