+++ src/components/ProductElements/ProductBlock/ProductBlock.jsx
@@ -1,15 +1,15 @@
 import './ProductBlock.scss';
-import Counter from '../../../components/UI/Counter/Counter';
-import { Button } from '../../../components/UI/Button/Button';
-import IconAvailable from '../../../components/UI/Icon/Icon_available';
-import IconNotAvailable from '../../../components/UI/Icon/Icon_not-available';
-import IconScales from '../../../components/UI/Icon/Icon_scales';
-import IconHearth from '../../../components/UI/Icon/Icon_hearth';
-import IconInfo from '../../../components/UI/Icon/Icon_info';
-import ProductRating from '../../../components/ProductRating/ProductRating';
-import CharacteristicColor from '../../../components/CharacteristicColor/CharacteristicColor';
+import Counter from '../../UI/Counter/Counter';
+import { Button } from '../../UI/Button/Button';
+import IconAvailable from '../../UI/Icon/Icon_available';
+import IconNotAvailable from '../../UI/Icon/Icon_not-available';
+import IconScales from '../../UI/Icon/Icon_scales';
+import IconHearth from '../../UI/Icon/Icon_hearth';
+import IconInfo from '../../UI/Icon/Icon_info';
+import ProductRating from '../ProductRating/ProductRating';
+import CharacteristicColor from '../CharacteristicColor/CharacteristicColor';
 import { useEffect, useState } from 'react';
-import CommentsBlock from '../../../components/CommentsBlock/CommentsBlock';
+import CommentsBlock from '../CommentsBlock/CommentsBlock';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import 'swiper/css';
 import { Navigation } from 'swiper/modules';
