+++ src/components/ProductElements/ProductCard/ProductCard.jsx
@@ -1,16 +1,15 @@
-import { Button } from '../UI/Button/Button';
-
-import IconInfo from '../UI/Icon/Icon_info';
-import IconHearth from '../UI/Icon/Icon_hearth';
-import IconScales from '../UI/Icon/Icon_scales';
-import SliderImage from '../SliderImage/SliderImage';
 import './ProductCard.scss';
+import { Button } from '../../UI/Button/Button';
+import IconInfo from '../../UI/Icon/Icon_info';
+import IconHearth from '../../UI/Icon/Icon_hearth';
+import IconScales from '../../UI/Icon/Icon_scales';
+import SliderImage from '../../SliderImage/SliderImage';
 import { Link } from 'react-router-dom';
-import Tooltip from '../UI/Tooltip/Tooltip';
-import IconInfoFil from '../UI/Icon/Icon_info_fill';
+import Tooltip from '../../UI/Tooltip/Tooltip';
+import IconInfoFil from '../../UI/Icon/Icon_info_fill';
 import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { addProduct, deleteProduct } from '../../store/slices/basketSlice';
+import { addProduct, deleteProduct } from '../../../store/slices/basketSlice';
 
 function ProductCard({ product }) {
   const [isProductSelect, setIsProductSelect] = useState(false);
