+++ src/components/ProductElements/ProductCard/ProductCard.scss
@@ -1,4 +1,4 @@
-@import '../../styles/global';
+@import '../../../styles/global';
 
 .card {
   width: $cardWidth;
