+++ src/components/ProductElements/ProductCardBasket/ProductCardBasket.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import './ProductCardBasket.scss';
-import Checkbox from '../UI/Checkbox/Checkbox';
-import IconTrash from '../UI/Icon/Icon_trash';
-import VendorCode from '../UI/VendorCode/VendorCode';
-import imageStub from '../../images/basket/Stub_132_128.jpg';
-import IconHearth from '../UI/Icon/Icon_hearth';
-import Counter from '../UI/Counter/Counter';
+import Checkbox from '../../UI/Checkbox/Checkbox';
+import IconTrash from '../../UI/Icon/Icon_trash';
+import VendorCode from '../../UI/VendorCode/VendorCode';
+import imageStub from '../../../images/basket/Stub_132_128.jpg';
+import IconHearth from '../../UI/Icon/Icon_hearth';
+import Counter from '../../UI/Counter/Counter';
 import ProductCharacteristicsList from '../ProductCharacteristic/ProductCharacteristic';
 import { useDispatch } from 'react-redux';
-import { changeQuantity, deleteProduct } from '../../store/slices/basketSlice.js';
+import { changeQuantity, deleteProduct } from '../../../store/slices/basketSlice.js';
 
 /**
  * Компонент ProductCardBasket для отображения товара.
