+++ src/components/ProductElements/ProductCardBasket/ProductCardBasket.scss
@@ -1,4 +1,4 @@
-@import '../../styles/global';
+@import '../../../styles/global';
 
 .product-card-basket {
   font-family: $fontFamilyNunitoSans;
