+++ src/components/ProductElements/ProductCardContainer/ProductCardContainer.scss
@@ -1,4 +1,4 @@
-@import '../../styles/global';
+@import '../../../styles/global';
 
 .cards-container {
   width: 100%;
