+++ src/components/ProductElements/ProductCharacteristic/ProductCharacteristic.scss
@@ -1,4 +1,4 @@
-@import '../../styles/global';
+@import '../../../styles/global';
 
 .product-characteristic {
   font-family: $fontFamilyNunitoSans;
