+++ src/components/ProductElements/ProductRating/ProductRating.scss
@@ -1,4 +1,4 @@
-@import '../../styles/global';
+@import '../../../styles/global';
 
 .product-rating {
   display: flex;
