+++ src/components/SliderImage/SliderImage.jsx
@@ -1,5 +1,6 @@
 import './SliderImage.scss';
 import Dot from './Dot/Dot';
+import noImage from '../../images/nophoto.png';
 
 export default function SliderImage({ images }) {
   const width = 100 / images.length;
@@ -11,9 +12,10 @@ export default function SliderImage({ images }) {
   const handleMouseEnter = (event) => {
     const card = getCurrentSlider(event);
     if (images.length !== 0) {
-      card.querySelector('.slider__images').style.transform = `translateX(${
-        -card.querySelector('.slider__image').width * event.target.id
-      }px)`;
+      card
+        .querySelector('.slider__images')
+        .style
+        .transform = `translateX(${-card.querySelector('.slider__image').width * event.target.id}px)`;
       [...card.querySelectorAll('.dot')].map((item) => item.classList.remove('dot_fill'));
       [...card.querySelectorAll('.dot')]
         .filter((item) => item.id === event.target.id)[0]
@@ -41,9 +43,11 @@ export default function SliderImage({ images }) {
   return (
     <div className="slider" onMouseLeave={handleStartState}>
       <div className="slider__images">
-        {images.map((image, idx) => (
-          <img key={idx} id={idx} className="slider__image" src={image.image} alt="" />
-        ))}
+        {images.length !== 0
+          ? (images.map((image, idx) => (
+            <img key={idx} id={idx} className="slider__image" src={image.image} alt="" />
+          )))
+          : (<img className="slider__image" src={noImage} alt="" />)}
       </div>
       <div className="slider__nav">
         {images.map((image, idx) => (
