+++ src/components/UI/ButtonScrollUp/ButtonScrollUp.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import './ButtonScrollUp.scss';
-import IconArrowUp from '../UI/Icon/Icon_arrow-up';
+import IconArrowUp from '../Icon/Icon_arrow-up';
 
 function ButtonScrollUp() {
   const [isVisible, setIsVisible] = useState(false);
