+++ src/components/UI/ButtonScrollUp/ButtonScrollUp.scss
@@ -1,4 +1,4 @@
-@import '../../styles/global';
+@import '../../../styles/global';
 
 .button-scroll-up {
   max-width: 1672px;
